rebases filters from main to View
