+++ src/view.js
@@ -1,12 +1,39 @@
 import {EventEmitter} from "./event-emitter";
+import {Filter} from './filter';
 import TripPoint from './trip-point';
 import TripPointDetailed from './trip-point-detailed';
+import moment from 'moment';
 
+const FILTERS_DATA = [{
+  textFilter: `Everything`,
+  doFilter() {
+    return true;
+  }
+},
+{
+  textFilter: `Future`,
+  doFilter(point) {
+    if (moment().isBefore(moment(point.date, `DD MMM`))) {
+      return true;
+    }
+    return false;
+  }
+},
+{
+  textFilter: `Past`,
+  doFilter(point) {
+    if (moment(point.date, `DD MMM`).isBefore(moment())) {
+      return true;
+    }
+    return false;
+  }
+}];
 class View extends EventEmitter {
-  constructor(model, container) {
+  constructor(model, pointContainer, filterContainer) {
     super();
     this._model = model;
-    this._container = container;
+    this._pointsContainer = pointContainer;
+    this._filterContainer = filterContainer;
     model.on(`pointsLoaded`, () => {
       return this.show();
     });
@@ -21,6 +48,7 @@ class View extends EventEmitter {
   show() {
     this._generateViews(this._model.points);
     this._renderPoints(this._arrayPoints);
+    this._addFilters();
   }
   _deletePoint(id) {
     this.emit(`deleted`, id);
@@ -35,7 +63,7 @@ class View extends EventEmitter {
 
   _renderPoints(array) {
     array.forEach((point) => {
-      this._container.appendChild(point.render());
+      this._pointsContainer.appendChild(point.render());
     });
   }
   _generateViews(data) {
@@ -45,12 +73,12 @@ class View extends EventEmitter {
         const tripPointDetailed = new TripPointDetailed(pointData, this._model, this);
         tripPoint.onClickPoint = () => {
           tripPointDetailed.render();
-          this._container.replaceChild(tripPointDetailed.element, tripPoint.element);
+          this._pointsContainer.replaceChild(tripPointDetailed.element, tripPoint.element);
           tripPoint.unrender();
         };
         tripPointDetailed.onClose = () => {
           tripPoint.render();
-          this._container.replaceChild(tripPoint.element, tripPointDetailed.element);
+          this._pointsContainer.replaceChild(tripPoint.element, tripPointDetailed.element);
           tripPointDetailed.unrender();
         };
         tripPointDetailed.onSaveClick = (newData) => {
@@ -66,6 +94,36 @@ class View extends EventEmitter {
       return null;
     });
   }
+
+  _clearPointsContainer() {
+    this._pointsContainer.innerHTML = ``;
+  }
+
+  _addFilters() {
+    for (let filterData of FILTERS_DATA) {
+      let filter = new Filter(filterData);
+      filter.onFilter = () => {
+        this._clearPointsContainer();
+        this._arrayPoints.filter((point) => {
+          return point !== null;
+        }).forEach((point) => {
+          if (point.element !== null) {
+            point.unrender();
+          }
+        });
+        this._renderPoints(this._arrayPoints.filter((point) => {
+          return point !== null;
+        }).filter((point) => {
+          return FILTERS_DATA.find((it) => {
+            return it === filterData;
+          }).doFilter(point);
+        }));
+      };
+      this._filterContainer.appendChild(filter.render());
+    }
+  }
+
+
 }
